# swagger.json
{"openapi":"3.0.0","info":{"title":"Sandbox Master API","version":"1.0.0","description":"API para gerenciamento de sandboxes Docker"},"servers":[{"url":"http://localhost:3000","description":"Servidor local"}],"components":{"schemas":{"PortMapping":{"type":"object","properties":{"hostPort":{"type":"number","example":16100},"internalPort":{"type":"number","example":3000}}},"BuildInfo":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"containerName":{"type":"string"},"imageName":{"type":"string"},"status":{"type":"string","enum":["building","running","stopped","error"]},"portMappings":{"type":"array","items":{"$ref":"#/components/schemas/PortMapping"}},"createdAt":{"type":"string","format":"date-time"}}},"ContainerInfo":{"type":"object","properties":{"name":{"type":"string"},"status":{"type":"string"},"image":{"type":"string"},"id":{"type":"string"},"portMappings":{"type":"array","items":{"$ref":"#/components/schemas/PortMapping"}},"type":{"type":"string","enum":["sub","build"]}}},"Error":{"type":"object","properties":{"error":{"type":"string"},"logs":{"type":"string"}}}}},"paths":{"/sandbox/build":{"post":{"summary":"Faz upload de um arquivo (tar/zip) com Dockerfile, constrói e executa","tags":["Builds"],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["file","name"],"properties":{"file":{"type":"string","format":"binary"},"name":{"type":"string","example":"meu-projeto"},"internalPorts":{"type":"string","description":"Array JSON de portas internas (ex: [3000])"},"env":{"type":"string","description":"Objeto JSON com variáveis de ambiente"}}}}}},"responses":{"200":{"description":"Build concluído e container em execução","content":{"application/json":{"schema":{"allOf":[{"type":"object","properties":{"message":{"type":"string"}}},{"$ref":"#/components/schemas/BuildInfo"}]}}}},"400":{"description":"Parâmetros inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/builds":{"get":{"summary":"Lista todos os builds ativos","tags":["Builds"],"responses":{"200":{"description":"Lista de builds","content":{"application/json":{"schema":{"type":"object","properties":{"builds":{"type":"array","items":{"$ref":"#/components/schemas/BuildInfo"}}}}}}}}}},"/sandbox/builds/{name}/stop":{"post":{"summary":"Para um container de build","tags":["Builds"],"parameters":[{"in":"path","name":"name","required":true,"schema":{"type":"string"},"description":"Nome base do build (sem prefixo)"}],"responses":{"200":{"description":"Build parado","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/builds/{name}":{"delete":{"summary":"Remove um container de build","tags":["Builds"],"parameters":[{"in":"path","name":"name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Build removido","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/cmd":{"post":{"summary":"Executa um comando no diretório sandbox","tags":["Comandos"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["command"],"properties":{"command":{"type":"string","example":"ls -la"}}}}}},"responses":{"200":{"description":"Saída do comando","content":{"application/json":{"schema":{"type":"object","properties":{"output":{"type":"string"}}}}}},"400":{"description":"Comando inválido","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/file":{"post":{"summary":"Cria um arquivo no diretório sandbox","tags":["Arquivos"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["filename"],"properties":{"filename":{"type":"string","example":"test.txt"},"content":{"type":"string","example":"Hello World"}}}}}},"responses":{"200":{"description":"Arquivo criado","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"path":{"type":"string"}}}}}},"400":{"description":"Nome do arquivo inválido","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/file/{filename}":{"delete":{"summary":"Remove um arquivo do diretório sandbox","tags":["Arquivos"],"parameters":[{"in":"path","name":"filename","required":true,"schema":{"type":"string"},"description":"Nome do arquivo"}],"responses":{"200":{"description":"Arquivo removido","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/install":{"post":{"summary":"Instala um pacote npm no diretório sandbox","tags":["Comandos"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["pkg"],"properties":{"pkg":{"type":"string","example":"express"}}}}}},"responses":{"200":{"description":"Saída da instalação","content":{"application/json":{"schema":{"type":"object","properties":{"output":{"type":"string"}}}}}},"400":{"description":"Nome do pacote inválido","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Erro interno","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/sandbox/restart":{"post":{"summary":"Reinicia o servidor master (sai do processo)","tags":["Sistema"],"responses":{"200":{"description":"Reiniciando...","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}}}}},"tags":[]}